added site_cookbooks

+++ cookbooks/java/libraries/default.rb
@@ -6,6 +6,10 @@ module JavaUtils
   end
 
   def do_copy(src='/etc/hosts', dest='/tmp1')
+    unless File.exists?(File.dirname(dest))
+      Chef::Log::error('Cant copy into non-existend destination ... quit')
+      return nil
+    end
     unless File.exists?(dest)
       FileUtils.copy(src, dest)
     end
